fix: fixes success page extraction and initial route
