+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:warren_first_task/portfolio/view/crypto_page.dart';
 
 import 'shared/routes/route_generator.dart';
 
@@ -9,7 +10,7 @@ void main() {
       child: MaterialApp(
         title: 'Crypto Listing',
         debugShowCheckedModeBanner: false,
-        initialRoute: '/crypto-page',
+        initialRoute: CryptoPage.route,
         onGenerateRoute: RouteGenerator.generateRoute,
       ),
     ),
